Remove SafeMath.sol and move tryXxx function to Math.sol
